+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Connector widget sets the default alias as "default" if the `.ini` file has no connections
 * Connector widget does not modify `.ini` file if the connection fails (#68)
 * Connector widget allows editing connections (#61)
+* Added Oracle, Microsoft SQLServer, Redshift in DB templates.
 
 ## 0.2.0